Fix Jenkinsfile so builds work again (#48)
* Fix Jenkinsfile so builds work again

This had shell syntax in the groovy file. Fix it so that it
embeds the shell syntax inside a Jenkins groovy `sh` call.

* Fix syntax: else vs elif

* brew install is not idempotent

This fails, see https://travis-ci.com/ModusCreateOrg/genus/jobs/153073183

This is why we have to have all the brew dependencies in the Brewfile,
and not try to `brew install` anything after that.

* Switch to ssh for git if creative-engine not found

* Force update of cmake if it is outdated with brew

* Ooops, revert Brewfile, that was committed in error

* Force update of cmake (brew is not idempotent)

* Crank up debugging

* Use script instead of install with travis

This was failing with:

```

$ ./scripts/build.sh
...  # successful build of genus !
0.01s$ ./configure && make && make test
/Users/travis/.travis/job_stages: line 104: ./configure: No such file or directory
The command "./configure && make && make test" exited with 127.
Done. Your build exited with 1.
```
It looks like the default Travis script is to run:
```./configure && make && make test```
So let's have it just run the script.

* Crank debug level back down to normal